Implement a version of rc log for swift.
