+++ components/rc_log/Cargo.toml
@@ -8,7 +8,13 @@ authors = ["Thom Chiovoloni <tchiovoloni@mozilla.com>"]
 name = "rc_log_ffi"
 crate-type = ["lib", "staticlib", "cdylib"]
 
+[features]
+default = []
+# Required for gradle in robolectric
+force_android = []
+
 [dependencies]
 log = { version = "0.4.6", features = ["std"] }
 ffi-support = { path = "../support/ffi" }
 lazy_static = "1.2.0"
+cfg-if = "0.1.6"
