+++ components/rc_log/android/build.gradle
@@ -76,6 +76,10 @@ cargo {
     profile = "release"
 
     exec = rootProject.ext.cargoExec
+
+    features {
+        defaultAnd("force_android")
+    }
 }
 
 configurations {
